backend/api/urls.py
from django.urls import path, include
from drf_spectacular.views import SpectacularAPIView, SpectacularSwaggerView
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView, TokenVerifyView

from api.routers import cat_router, rev_router, item_router, payment_router
from api.views import health

app_name = 'api'

urlpatterns = [
    path('', include(item_router.urls)),
    path('', include(cat_router.urls)),
    path('', include(rev_router.urls)),
    path('', include(payment_router.urls)),
    path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('token/verify/', TokenVerifyView.as_view(), name='token_verify'),
    path('schema/', SpectacularAPIView.as_view(), name='schema'),
    path('docs/', SpectacularSwaggerView.as_view(url_name='api:schema'), name='docs'),
    path('health/', health, name='helth'),
]